+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Added ptv vissim and carla co-simulation
   * API extensions:
     - Added new methods to `Map`: `get_all_landmarks`, `get_all_landmarks_from_id` and `get_all_landmarks_of_type`
+  * Added synchronization of traffic lights in sumo co-simulation
 
 ## CARLA 0.9.8